+++ tests/PluralTranslationTest.php
@@ -6,9 +6,7 @@ use Maartenpaauw\Translation\PluralTranslation;
 
 it('should be possible to generate singular and plural combined translation messages', function () {
     // Act
-    $message = (new PluralTranslation())
-        ->singular('There is one apple')
-        ->plural('There are many apples');
+    $message = new PluralTranslation('There is one apple', 'There are many apples');
 
     // Assert
     expect($message)->toEqual('There is one apple|There are many apples');